Improve exception handling in some unit tests.
Remove unnecessary function throws (a copy and paste mistake).

Attempt to work around SonarQubes incorrect assumption that multiple exceptions can be thrown in the test simply because a `RuntimeException` is being created and passed as an argument for a unit test.
Relocate the `RuntimeException` declaration to a static final at the class level.