Fix all cases of static final vs final static ordering and updated recent tests.
I seem to have a habit of typing `final static` given the impact of the word `final` is stronger in my mind.
However, the proper order per SonarQube is `static final`.
I searched the entire project and found additional areas that needed this fix.

I missed the `EmailDelegateAddressFailureTest` test in the previous commit.

Remove the double `WorksWorks` in a test name.